feat: update devnet struct; add devnet example
